+++ angel_orm_generator/test/models/book.g.dart
@@ -3,6 +3,142 @@
 part of angel_orm.generator.models.book;
 
 // **************************************************************************
+// OrmGenerator
+// **************************************************************************
+
+class BookQuery extends Query<Book, BookQueryWhere> {
+  @override
+  final BookQueryValues values = new BookQueryValues();
+
+  @override
+  final BookQueryWhere where = new BookQueryWhere();
+
+  @override
+  get tableName {
+    return 'books';
+  }
+
+  @override
+  get fields {
+    return const [
+      'id',
+      'author_id',
+      'partner_author_id',
+      'name',
+      'created_at',
+      'updated_at'
+    ];
+  }
+
+  @override
+  BookQueryWhere newWhereClause() {
+    return new BookQueryWhere();
+  }
+
+  static Book parseRow(List row) {
+    var model = new Book(
+        id: row[0].toString(),
+        name: (row[3] as String),
+        createdAt: (row[4] as DateTime),
+        updatedAt: (row[5] as DateTime));
+    if (row.length > 6) {
+      model =
+          model.copyWith(author: AuthorQuery.parseRow(row.skip(6).toList()));
+    }
+    if (row.length > 10) {
+      model = model.copyWith(
+          partnerAuthor: AuthorQuery.parseRow(row.skip(10).toList()));
+    }
+    return model;
+  }
+
+  @override
+  deserialize(List row) {
+    return parseRow(row);
+  }
+
+  @override
+  get(executor) {
+    leftJoin('authors', 'author_id', 'id',
+        additionalFields: const ['name', 'created_at', 'updated_at']);
+    leftJoin('authors', 'partner_author_id', 'id',
+        additionalFields: const ['name', 'created_at', 'updated_at']);
+    return super.get(executor);
+  }
+}
+
+class BookQueryWhere extends QueryWhere {
+  final NumericSqlExpressionBuilder<int> id =
+      new NumericSqlExpressionBuilder<int>('id');
+
+  final NumericSqlExpressionBuilder<int> authorId =
+      new NumericSqlExpressionBuilder<int>('author_id');
+
+  final NumericSqlExpressionBuilder<int> partnerAuthorId =
+      new NumericSqlExpressionBuilder<int>('partner_author_id');
+
+  final StringSqlExpressionBuilder name =
+      new StringSqlExpressionBuilder('name');
+
+  final DateTimeSqlExpressionBuilder createdAt =
+      new DateTimeSqlExpressionBuilder('created_at');
+
+  final DateTimeSqlExpressionBuilder updatedAt =
+      new DateTimeSqlExpressionBuilder('updated_at');
+
+  @override
+  get expressionBuilders {
+    return [id, authorId, partnerAuthorId, name, createdAt, updatedAt];
+  }
+}
+
+class BookQueryValues extends MapQueryValues {
+  int get id {
+    return (values['id'] as int);
+  }
+
+  void set id(int value) => values['id'] = value;
+  int get authorId {
+    return (values['author_id'] as int);
+  }
+
+  void set authorId(int value) => values['author_id'] = value;
+  int get partnerAuthorId {
+    return (values['partner_author_id'] as int);
+  }
+
+  void set partnerAuthorId(int value) => values['partner_author_id'] = value;
+  String get name {
+    return (values['name'] as String);
+  }
+
+  void set name(String value) => values['name'] = value;
+  DateTime get createdAt {
+    return (values['created_at'] as DateTime);
+  }
+
+  void set createdAt(DateTime value) => values['created_at'] = value;
+  DateTime get updatedAt {
+    return (values['updated_at'] as DateTime);
+  }
+
+  void set updatedAt(DateTime value) => values['updated_at'] = value;
+  void copyFrom(Book model) {
+    values.addAll({
+      'name': model.name,
+      'created_at': model.createdAt,
+      'updated_at': model.updatedAt
+    });
+    if (model.author != null) {
+      values['author_id'] = int.parse(model.author.id);
+    }
+    if (model.partnerAuthor != null) {
+      values['partner_author_id'] = int.parse(model.partnerAuthor.id);
+    }
+  }
+}
+
+// **************************************************************************
 // JsonModelGenerator
 // **************************************************************************
 
@@ -12,7 +148,6 @@ class Book extends _Book {
       {this.id,
       this.author,
       this.partnerAuthor,
-      this.authorId,
       this.name,
       this.createdAt,
       this.updatedAt});
@@ -27,9 +162,6 @@ class Book extends _Book {
   final Author partnerAuthor;
 
   @override
-  final int authorId;
-
-  @override
   final String name;
 
   @override
@@ -42,7 +174,6 @@ class Book extends _Book {
       {String id,
       Author author,
       Author partnerAuthor,
-      int authorId,
       String name,
       DateTime createdAt,
       DateTime updatedAt}) {
@@ -50,7 +181,6 @@ class Book extends _Book {
         id: id ?? this.id,
         author: author ?? this.author,
         partnerAuthor: partnerAuthor ?? this.partnerAuthor,
-        authorId: authorId ?? this.authorId,
         name: name ?? this.name,
         createdAt: createdAt ?? this.createdAt,
         updatedAt: updatedAt ?? this.updatedAt);
@@ -61,7 +191,6 @@ class Book extends _Book {
         other.id == id &&
         other.author == author &&
         other.partnerAuthor == partnerAuthor &&
-        other.authorId == authorId &&
         other.name == name &&
         other.createdAt == createdAt &&
         other.updatedAt == updatedAt;
@@ -69,8 +198,7 @@ class Book extends _Book {
 
   @override
   int get hashCode {
-    return hashObjects(
-        [id, author, partnerAuthor, authorId, name, createdAt, updatedAt]);
+    return hashObjects([id, author, partnerAuthor, name, createdAt, updatedAt]);
   }
 
   Map<String, dynamic> toJson() {